Close all connections on exit (Phase 1)
When an example application exits, iterate over all client
and server connections and send a CONNECTION_CLOSE message
on each before cleaning up all resources.

This is only a simple initial implementation (Phase 1)
that will successfully close a small number of connections.
Phase 2 will deal with larger numbers of connections and
the load that closing these will place on a single UDP
socket.