Rewrites ngtcp2_ksl
This change rewrites most of the code in ngtcp2_ksl and removes the
need for inf_key.  It also removes the restriction that the key must
exist in descendants.  ngtcp2_ksl now embeds arbitrary sized key in
ngtcp2_node.

ngtcp2_ksl is extended so that it can work with ngtcp2_range key.
ngtcp2_psl is removed in favor of ngtcp2_ksl.