Make sure that a packet which elicits key update confirmation is sent
This change fixes the bug that a packet which elicits key update
confirmation from the remote endpoint might not be sent.  This might
lead to connection failure if the remote endpoint keeps changing keys
very frequently.  Changing keys in this frequency is unusual, but it
is allowed by the spec, and we should deal with it.