Server other_versions defaults to v1
This commit makes server other_versions default to v1.  server VN is
adjusted to reflect this and now it sends only v1 (+ a reserved
version) in VN.  The side effect of this change is that we cannot
specify empty other_versions field.  But since other_versions cannot
be empty on incompatible version negotiation, we almost always need
non-empty server other_versions.