Do not return NGTCP2_ERR_STREAM_NOT_FOUND
Do not return NGTCP2_ERR_STREAM_NOT_FOUND from the following
functions:

- ngtcp2_conn_shutdown_stream
- ngtcp2_conn_shutdown_stream_write
- ngtcp2_conn_shutdown_stream_read
- ngtcp2_conn_extend_max_stream_offset

and returns 0 if a stream is not found.

The reason is that it is too tedious to check
NGTCP2_ERR_STREAM_NOT_FOUND in every invocation for these functions.
In almost all cases, if NGTCP2_ERR_STREAM_NOT_FOUND is returned, we
just simply ignores it and pretend that the function succeeds.